List filters (#162)
* add filters

* styles

* fix tests

* tests

* llint

* dedupe tests
